refactor(selector4nix): define `NarFile` and `NarFileKey` in separate modules
