+++ src/domain/nar_file/model/mod.rs
@@ -1,5 +1,7 @@
 mod nar_file;
+mod nar_file_key;
 mod nar_file_location;
 
-pub use nar_file::{NarFile, NarFileKey};
+pub use nar_file::NarFile;
+pub use nar_file_key::NarFileKey;
 pub use nar_file_location::NarFileLocation;
